+++ elasticsearch5/src/main/java/org/vertexium/elasticsearch5/ElasticsearchSearchIndexConfiguration.java
@@ -49,6 +49,8 @@ public class ElasticsearchSearchIndexConfiguration {
     private static final boolean FORCE_DISABLE_VERTEXIUM_PLUGIN_DEFAULT = false;
     public static final String INDEX_MAPPING_TOTAL_FIELDS_LIMIT = "indexMappingTotalFieldsLimit";
     public static final int INDEX_MAPPING_TOTAL_FIELDS_LIMIT_DEFAULT = 100000;
+    public static final String INDEX_REFRESH_INTERVAL = "indexRefreshInterval";
+    public static final Integer INDEX_REFRESH_INTERVAL_DEFAULT = null;
     public static final String PROPERTY_NAME_VISIBILITIES_STORE = "propertyNameVisibilitiesStore";
     public static final Class<? extends PropertyNameVisibilitiesStore> PROPERTY_NAME_VISIBILITIES_STORE_DEFAULT = MetadataTablePropertyNameVisibilitiesStore.class;
     public static final String EXCEPTION_HANDLER = "exceptionHandler";
@@ -127,6 +129,10 @@ public class ElasticsearchSearchIndexConfiguration {
 
     public int getIndexMappingTotalFieldsLimit() {
         return graphConfiguration.getInt(GraphConfiguration.SEARCH_INDEX_PROP_PREFIX + "." + INDEX_MAPPING_TOTAL_FIELDS_LIMIT, INDEX_MAPPING_TOTAL_FIELDS_LIMIT_DEFAULT);
+    }
+
+    public Integer getIndexRefreshInterval() {
+        return graphConfiguration.getInteger(GraphConfiguration.SEARCH_INDEX_PROP_PREFIX + "." + INDEX_REFRESH_INTERVAL, INDEX_REFRESH_INTERVAL_DEFAULT);
     }
 
     public int getTermAggregationShardSize() {